Converts hue into a component/platform solution
Spilts out hue hub into its own component which fetches the URL for API
requests. This is likely a work in progress.
